fix build

+++ features/transaction/buyerorder/src/main/java/com/tokopedia/buyerorder/unifiedhistory/list/view/fragment/UohListFragment.kt
@@ -50,7 +50,6 @@ import com.tokopedia.atc_common.data.model.request.AddToCartRequestParams
 import com.tokopedia.atc_common.domain.model.request.AddToCartMultiParam
 import com.tokopedia.buyerorder.R
 import com.tokopedia.buyerorder.common.util.BuyerConsts.ACTION_FINISH_ORDER
-import com.tokopedia.buyerorder.list.view.activity.OrderListActivity
 import com.tokopedia.buyerorder.unifiedhistory.common.util.UohConsts
 import com.tokopedia.buyerorder.unifiedhistory.common.util.UohConsts.ALL_DATE
 import com.tokopedia.buyerorder.unifiedhistory.common.util.UohConsts.ALL_PRODUCTS
@@ -1778,7 +1777,7 @@ class UohListFragment: BaseDaggerFragment(), RefreshHandler.OnRefreshHandlerList
                 listParamAtcMulti.add(AddToCartMultiParam(
                         productId = objParam.get(PRODUCT_ID).asLong,
                         productName = objParam.get(PRODUCT_NAME).asString,
-                        productPrice = objParam.get(PRODUCT_PRICE).asInt,
+                        productPrice = objParam.get(PRODUCT_PRICE).asLong,
                         qty = objParam.get(QUANTITY).asInt,
                         notes = objParam.get(NOTES).asString,
                         shopId = objParam.get(SHOP_ID).asInt,
